\name{cnPopData}
\docType{data}
\alias{cnPopData}
\title{Test data for the China border Group}
\description{
This dataset contains the 2014 population and average income per person in each of the 
China Areas  }
\usage{data(cnPopData)}
\format{
  A data frame with 18 observations, 1 for each Seer area, 
  on the following 12 variables.
  \describe{
    \item{area}{a character vector containing the China area full names.}
    \item{pop2013}{a numeric vector of the number of the county's population}
  }
}
\details{
 This dataset was pulled from the China ??? government website on Dec. 2014.
 It contains the population and average income per person for each of Kansas' 105 
 counties.}
\keyword{datasets}