\name{KansPopInc}
\docType{data}
\alias{KansPopInc}
\title{Test data for the Kansas border Group}
\description{
This dataset contains the 2014 population and average income per person in each of the 
105 Kansas counties.
  }
\usage{data(KansPopInc)}
\format{
  A data frame with 18 observations, 1 for each Seer area, 
  on the following 12 variables.
  \describe{
    \item{County}{a character vector containing the Kansas County Name.}
    \item{Pop}{a numeric vector of the number of the county's population}
    \item{AvgInc}{a numeric vector of the average person's income for the county.}
  }
}
\details{
 This dataset was pulled from the Kansas government website on Dec. 2014.
 It contains the population and average income per person for each of Kansas' 105 
 counties.}
\author{Linda W. Pickle and Jim Pearson of StatNet Consulting, LLC, Gaithersburg, MD}
\keyword{datasets}