\name{mdPopData}
\docType{data}
\alias{mdPopData}
\title{Test data for the Maryland border Group}
\description{
This dataset contains the county populations for 1970, 1980, 1990, 2000, and 2010 and 
the projected estimated populations for 2015, 2020, 2025, 2030, 2035 and 2040 for 
each Maryland county and Baltimore City.
  }
\usage{data(mdPopData)}
\format{
  A data frame with 24 rows, 1 for each county, with 12 variables for each county: 
  \describe{
    \item{Abbr}{a character vector abbreviation for the county.}
    \item{County}{a character vector containing the Maryland county name and Baltimore City.}
    \item{X1970}{a numeric vector of the county's population in 1970.}
    \item{X1980}{a numeric vector of the county's population in 1980.}
    \item{X1990}{a numeric vector of the county's population in 1990.}
    \item{X2000}{a numeric vector of the county's population in 2000.}
    \item{X2010}{a numeric vector of the county's population in 2010.}
    \item{X2015}{a numeric vector of the county's estimated population in 2015.}
    \item{X2020}{a numeric vector of the county's estimated population in 2020.}
    \item{X2025}{a numeric vector of the county's estimated population in 2025.}
    \item{X2030}{a numeric vector of the county's estimated population in 2030.}
    \item{X2035}{a numeric vector of the county's estimated population in 2035.}
    \item{X2040}{a numeric vector of the county's estimated population in 2000.}
 . }
}
\details{
 This dataset was pulled from the Maryland government website in January, 2015.
 It contains the population and estimated population for each of Maryland 
 counties and Baltimore City for 1970, 1980, 1990, 2000, and 2010 and estimated populations for years 2015, 
 2020, 2025, 2030, 2035, and 2040.
 }
\keyword{datasets}