\name{nyPopData}
\docType{data}
\alias{nyPopData}
\title{Test data for the New York border Group}
\description{
This dataset contains the 2014 population and average income per person in each of the 
New York counties.
  }
\usage{data(nyPopData)}
\format{
  A data frame with 63 rows, 1 for each county and 14 variables per county: 
  \describe{
    \item{Area}{a character vector containing the New York county and buorgh name.}
    \item{Est_2000}{a numeric vector of the county's estimated population for 2000.}
    \item{Est_2001}{a numeric vector of the county's estimated population for 2001.}
    \item{Est_2002}{a numeric vector of the county's estimated population for 2002.}
    \item{Est_2003}{a numeric vector of the county's estimated population for 2003.}
    \item{Est_2004}{a numeric vector of the county's estimated population for 2004.}
    \item{Est_2005}{a numeric vector of the county's estimated population for 2005.}
    \item{Est_2006}{a numeric vector of the county's estimated population for 2006.}
    \item{Est_2007}{a numeric vector of the county's estimated population for 2007.}
    \item{Est_2008}{a numeric vector of the county's estimated population for 2008.}
    \item{Est_2009}{a numeric vector of the county's estimated population for 2009.}
    \item{Est_2010}{a numeric vector of the county's estimated population for 2010.}
    \item{Pop_2000}{a numeric vector of the county's actual population for 2000.}
    \item{Pop_2010}{a numeric vector of the county's actual population for 2010.}
  }
}
\details{
 This dataset was pulled from the New York government website in January, 2015.
 It contains the actual and estimate populations of each county from 2000 to 2010.
 }
\keyword{datasets}