\name{Seer18Area}
\docType{data}
\alias{Seer18Area}
\title{Test data for 18 U.S. Seer Registries (of 20) for general dot, arrow, and bar glyphics.}
\description{
Randomly generated statistics to create rate, count and population data for the
18 U.S. Seer registries for testing of micromapSEER or micromapST using the "USSeerBG" border group. 
The data was generated by the SeerStat program and exported as a CSV file.  
The non-data lines at the bottom
of the file must be removed.
  }
\usage{data(Seer18Area)}
\format{
  A data frame with 18 observations, 1 for each Seer registries (sub-areas), 
  on the following 11 variables.
  \describe{
    \item{Registry}{a character vector of the name of the Seer Registry}
    \item{Rate20022006}{a numeric vector of the incident rates from 2002 to 2006}
    \item{Count20022006}{a numeric vector of the incident counts}
    \item{Pop20022006}{a numeric vector of the population}
    \item{Rate20072011}{a numeric vector of the incident rates from 2007 to 2011}
    \item{Count20072011}{a numeric vector of the incident counts}
    \item{Pop20072011}{a numeric vector of the population}
    \item{RateTrendAPC}{a numeric vector of the rate trends per Seer registry}
    \item{pValue}{a numeric vector of the P Value related to the Rate Trend}
    \item{LowerCI}{a numeric vector of the lower confidence interval}
    \item{UpperCI}{a numeric vector of the upper confidence interval}
  }
}
\details{
 This dataset is a randomly generated collection of data from 2002 to 2011 Seer Registry
 data to support testing of micromapST functions for the 18 U.S. Seer registries.  
 The data has no relationship to the  Seer registries and is only for test purposes.
 The row names are the Seer registry abbreviations.   
 This dataset is used in the micromapST examples.
}
\author{Linda W. Pickle and Jim Pearson of StatNet Consulting, LLC, Gaithersburg, MD}
\keyword{datasets}