\name{SeoulPopData}
\docType{data}
\alias{SeoulPopData}
\title{Test data for the Seoul South Korea city district border Group}
\description{
This dataset contains the 2012 population and area data for each of the 
districts in city of Seoul. }
\usage{data(SeoulPopData)}
\format{
  A data frame with 25 observations, 1 for each district, 
  on the following 5 variables.
  \describe{
    \item{District}{a character vector containing the full name of the Seoul district.}
    \item{City}{a character vector containing the city name => Seoul.}
    \item{Pop.2012}{a numeric vector of the number of the district's 2012 population.}
    \item{Area}{a numeric vector of the area square kilometers for the district.}
    \item{Founded}{a character vector containing the dates each district was founded.}
  }
}
\details{
 This dataset was pulled from the China ??? government website on Dec. 2014.
 It contains the population and average income per person for each of Kansas' 105 
 counties.}
\keyword{datasets}