\name{UKIrelandPopData}
\docType{data}
\alias{UKIrelandPopData}
\alias{UKIrelandPopData2}
\title{Test data for the UK-Ireland border Group}
\description{
The UKIrelandPopData and UKIrelandPopData2 datasets contain the county populations and area statistics for the UK and Ireland..
 }
\usage{data(UKIrelandPopData)}
\format{
  A data frame with 218 rows, 1 for each county, with 5 variables for each county: 
  \describe{
    \item{Name}{a character vector containing the UK-Ireland county name.}
    \item{Abbr}{a character vector containing the ISO 2 character abbreviation for the UK and Ireland country. 
    Warning: the UKIreland border group uses the newer ISO 3 character notation for it\'s abbreviations. }
    \item{Pop}{a numeric vector of the county's population.}
    \item{Area.M}{a numeric vector of the county's area in square miles.}
    \item{Area.Km}{a numeric vector of the county's area in square kilometers.}
  }
}
\details{
   This dataset was pulled from several UK and Ireland website in January, 2015.
   The difference between UKIrelandPopData and UKIrelandPopData2 is incorrect labeling of 
   Anglesey as Anglesey, Isle of. 
 }
\keyword{datasets}
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  
  