\name{UtahPopData}
\docType{data}
\alias{UtahPopData}
\title{Test data for the Utah state border Group}
\description{
This dataset contains the Utah county populations for each year from 1940 to 2011.
  }
\usage{data(UtahPopData)}
\format{
  A data frame with 29 rows, 1 for each county, with 73 variables for each county: 
  \describe{
    \item{County}{a character vector containing the Utah county full name.}
    \item{X1940}{a numeric vector of the county's population in 1940.}
    \item{X1941}{a numeric vector of the county's population in 1941.}
    \item{X1942}{a numeric vector of the county's population in 1942.}
    \item{X1943}{a numeric vector of the county's population in 1943.}
    \item{X1944}{a numeric vector of the county's population in 1944.}
    \item{X1945}{a numeric vector of the county's population in 1945.}
    \item{X1946}{a numeric vector of the county's population in 1946.}
    \item{X1947}{a numeric vector of the county's population in 1947.}
    \item{X1948}{a numeric vector of the county's population in 1948.}
    \item{X1949}{a numeric vector of the county's population in 1949.}
    \item{X1950}{a numeric vector of the county's population in 1950.}
    \item{X1951}{a numeric vector of the county's population in 1951.}
    \item{X1952}{a numeric vector of the county's population in 1952.}
    \item{X1953}{a numeric vector of the county's population in 1953.}
    \item{X1954}{a numeric vector of the county's population in 1954.}
    \item{X1955}{a numeric vector of the county's population in 1955.}
    \item{X1956}{a numeric vector of the county's population in 1956.}
    \item{X1957}{a numeric vector of the county's population in 1957.}
    \item{X1958}{a numeric vector of the county's population in 1958.}
    \item{X1959}{a numeric vector of the county's population in 1959.}
    \item{X1960}{a numeric vector of the county's population in 1960.}
    \item{X1961}{a numeric vector of the county's population in 1961.}
    \item{X1962}{a numeric vector of the county's population in 1962.}
    \item{X1963}{a numeric vector of the county's population in 1963.}
    \item{X1964}{a numeric vector of the county's population in 1964.}
    \item{X1965}{a numeric vector of the county's population in 1965.}
    \item{X1966}{a numeric vector of the county's population in 1966.}
    \item{X1967}{a numeric vector of the county's population in 1967.}
    \item{X1968}{a numeric vector of the county's population in 1968.}
    \item{X1969}{a numeric vector of the county's population in 1969.}
    \item{X1970}{a numeric vector of the county's population in 1970.}
    \item{X1971}{a numeric vector of the county's population in 1971.}
    \item{X1972}{a numeric vector of the county's population in 1972.}
    \item{X1973}{a numeric vector of the county's population in 1973.}
    \item{X1974}{a numeric vector of the county's population in 1974.}
    \item{X1975}{a numeric vector of the county's population in 1975.}
    \item{X1976}{a numeric vector of the county's population in 1976.}
    \item{X1977}{a numeric vector of the county's population in 1977.}
    \item{X1978}{a numeric vector of the county's population in 1978.}
    \item{X1979}{a numeric vector of the county's population in 1979.}
    \item{X1980}{a numeric vector of the county's population in 1980.}
    \item{X1981}{a numeric vector of the county's population in 1981.}
    \item{X1982}{a numeric vector of the county's population in 1982.}
    \item{X1983}{a numeric vector of the county's population in 1983.}
    \item{X1984}{a numeric vector of the county's population in 1984.}
    \item{X1985}{a numeric vector of the county's population in 1985.}
    \item{X1986}{a numeric vector of the county's population in 1986.}
    \item{X1987}{a numeric vector of the county's population in 1987.}
    \item{X1988}{a numeric vector of the county's population in 1988.}
    \item{X1989}{a numeric vector of the county's population in 1989.}
    \item{X1990}{a numeric vector of the county's population in 1990.}
    \item{X1991}{a numeric vector of the county's population in 1991.}
    \item{X1992}{a numeric vector of the county's population in 1992.}
    \item{X1993}{a numeric vector of the county's population in 1993.}
    \item{X1994}{a numeric vector of the county's population in 1994.}
    \item{X1995}{a numeric vector of the county's population in 1995.}
    \item{X1996}{a numeric vector of the county's population in 1996.}
    \item{X1997}{a numeric vector of the county's population in 1997.}
    \item{X1998}{a numeric vector of the county's population in 1998.}
    \item{X1999}{a numeric vector of the county's population in 1999.}
    \item{X2000}{a numeric vector of the county's population in 2000.}
    \item{X2001}{a numeric vector of the county's population in 2001.}
    \item{X2002}{a numeric vector of the county's population in 2002.}
    \item{X2003}{a numeric vector of the county's population in 2003.}
    \item{X2004}{a numeric vector of the county's population in 2004.}
    \item{X2005}{a numeric vector of the county's population in 2005.}
    \item{X2006}{a numeric vector of the county's population in 2006.}
    \item{X2007}{a numeric vector of the county's population in 2007.}
    \item{X2008}{a numeric vector of the county's population in 2008.}
    \item{X2009}{a numeric vector of the county's population in 2009.}
    \item{X2010}{a numeric vector of the county's population in 2010.}
    \item{X2011}{a numeric vector of the county's population in 2011.}
 . }
}
\details{
 This dataset was pulled from the Utah government website in January, 2015.
 }
\keyword{datasets}